+++ accounting/src/apideck_accounting_unify/ageddebtors_sdk.py
@@ -98,6 +98,7 @@ class AgedDebtorsSDK(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.agedDebtorsOne",
                 oauth2_scopes=[],
@@ -259,6 +260,7 @@ class AgedDebtorsSDK(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.agedDebtorsOne",
                 oauth2_scopes=[],