+++ accounting/src/apideck_accounting_unify/attachments.py
@@ -108,6 +108,7 @@ class Attachments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsAll",
                 oauth2_scopes=[],
@@ -297,6 +298,7 @@ class Attachments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsAll",
                 oauth2_scopes=[],
@@ -490,6 +492,7 @@ class Attachments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsUpload",
                 oauth2_scopes=[],
@@ -657,6 +660,7 @@ class Attachments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsUpload",
                 oauth2_scopes=[],
@@ -817,6 +821,7 @@ class Attachments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsOne",
                 oauth2_scopes=[],
@@ -977,6 +982,7 @@ class Attachments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsOne",
                 oauth2_scopes=[],
@@ -1134,6 +1140,7 @@ class Attachments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsDelete",
                 oauth2_scopes=[],
@@ -1291,6 +1298,7 @@ class Attachments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsDelete",
                 oauth2_scopes=[],
@@ -1452,6 +1460,7 @@ class Attachments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsDownload",
                 oauth2_scopes=[],
@@ -1618,6 +1627,7 @@ class Attachments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.attachmentsDownload",
                 oauth2_scopes=[],