+++ accounting/src/apideck_accounting_unify/bankfeedaccounts.py
@@ -98,6 +98,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsAll",
                 oauth2_scopes=[],
@@ -283,6 +284,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsAll",
                 oauth2_scopes=[],
@@ -496,6 +498,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsAdd",
                 oauth2_scopes=[],
@@ -684,6 +687,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsAdd",
                 oauth2_scopes=[],
@@ -838,6 +842,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsOne",
                 oauth2_scopes=[],
@@ -992,6 +997,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsOne",
                 oauth2_scopes=[],
@@ -1183,6 +1189,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsUpdate",
                 oauth2_scopes=[],
@@ -1374,6 +1381,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsUpdate",
                 oauth2_scopes=[],
@@ -1525,6 +1533,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsDelete",
                 oauth2_scopes=[],
@@ -1676,6 +1685,7 @@ class BankFeedAccounts(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedAccountsDelete",
                 oauth2_scopes=[],