+++ accounting/src/apideck_accounting_unify/bankfeedstatements.py
@@ -99,6 +99,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsAll",
                 oauth2_scopes=[],
@@ -284,6 +285,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsAll",
                 oauth2_scopes=[],
@@ -497,6 +499,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsAdd",
                 oauth2_scopes=[],
@@ -685,6 +688,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsAdd",
                 oauth2_scopes=[],
@@ -839,6 +843,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsOne",
                 oauth2_scopes=[],
@@ -993,6 +998,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsOne",
                 oauth2_scopes=[],
@@ -1184,6 +1190,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsUpdate",
                 oauth2_scopes=[],
@@ -1375,6 +1382,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsUpdate",
                 oauth2_scopes=[],
@@ -1526,6 +1534,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsDelete",
                 oauth2_scopes=[],
@@ -1677,6 +1686,7 @@ class BankFeedStatements(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.bankFeedStatementsDelete",
                 oauth2_scopes=[],