+++ accounting/src/apideck_accounting_unify/billpayments.py
@@ -107,6 +107,7 @@ class BillPayments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsAll",
                 oauth2_scopes=[],
@@ -302,6 +303,7 @@ class BillPayments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsAll",
                 oauth2_scopes=[],
@@ -582,6 +584,7 @@ class BillPayments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsAdd",
                 oauth2_scopes=[],
@@ -835,6 +838,7 @@ class BillPayments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsAdd",
                 oauth2_scopes=[],
@@ -989,6 +993,7 @@ class BillPayments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsOne",
                 oauth2_scopes=[],
@@ -1143,6 +1148,7 @@ class BillPayments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsOne",
                 oauth2_scopes=[],
@@ -1399,6 +1405,7 @@ class BillPayments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsUpdate",
                 oauth2_scopes=[],
@@ -1655,6 +1662,7 @@ class BillPayments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsUpdate",
                 oauth2_scopes=[],
@@ -1806,6 +1814,7 @@ class BillPayments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsDelete",
                 oauth2_scopes=[],
@@ -1957,6 +1966,7 @@ class BillPayments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billPaymentsDelete",
                 oauth2_scopes=[],