+++ accounting/src/apideck_accounting_unify/bills.py
@@ -107,6 +107,7 @@ class Bills(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsAll",
                 oauth2_scopes=[],
@@ -302,6 +303,7 @@ class Bills(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsAll",
                 oauth2_scopes=[],
@@ -621,6 +623,7 @@ class Bills(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsAdd",
                 oauth2_scopes=[],
@@ -913,6 +916,7 @@ class Bills(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsAdd",
                 oauth2_scopes=[],
@@ -1067,6 +1071,7 @@ class Bills(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsOne",
                 oauth2_scopes=[],
@@ -1221,6 +1226,7 @@ class Bills(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsOne",
                 oauth2_scopes=[],
@@ -1516,6 +1522,7 @@ class Bills(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsUpdate",
                 oauth2_scopes=[],
@@ -1811,6 +1818,7 @@ class Bills(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsUpdate",
                 oauth2_scopes=[],
@@ -1962,6 +1970,7 @@ class Bills(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsDelete",
                 oauth2_scopes=[],
@@ -2113,6 +2122,7 @@ class Bills(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.billsDelete",
                 oauth2_scopes=[],