+++ accounting/src/apideck_accounting_unify/companyinfo_sdk.py
@@ -88,6 +88,7 @@ class CompanyInfoSDK(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.companyInfoOne",
                 oauth2_scopes=[],
@@ -239,6 +240,7 @@ class CompanyInfoSDK(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.companyInfoOne",
                 oauth2_scopes=[],