+++ accounting/src/apideck_accounting_unify/connectionsettings.py
@@ -89,6 +89,7 @@ class ConnectionSettings(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.connectionSettingsAll",
                 oauth2_scopes=[],
@@ -241,6 +242,7 @@ class ConnectionSettings(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.connectionSettingsAll",
                 oauth2_scopes=[],
@@ -426,6 +428,7 @@ class ConnectionSettings(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.connectionSettingsUpdate",
                 oauth2_scopes=[],
@@ -611,6 +614,7 @@ class ConnectionSettings(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.connectionSettingsUpdate",
                 oauth2_scopes=[],