+++ accounting/src/apideck_accounting_unify/creditnotes.py
@@ -111,6 +111,7 @@ class CreditNotes(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesAll",
                 oauth2_scopes=[],
@@ -310,6 +311,7 @@ class CreditNotes(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesAll",
                 oauth2_scopes=[],
@@ -620,6 +622,7 @@ class CreditNotes(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesAdd",
                 oauth2_scopes=[],
@@ -903,6 +906,7 @@ class CreditNotes(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesAdd",
                 oauth2_scopes=[],
@@ -1057,6 +1061,7 @@ class CreditNotes(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesOne",
                 oauth2_scopes=[],
@@ -1211,6 +1216,7 @@ class CreditNotes(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesOne",
                 oauth2_scopes=[],
@@ -1497,6 +1503,7 @@ class CreditNotes(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesUpdate",
                 oauth2_scopes=[],
@@ -1783,6 +1790,7 @@ class CreditNotes(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesUpdate",
                 oauth2_scopes=[],
@@ -1934,6 +1942,7 @@ class CreditNotes(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesDelete",
                 oauth2_scopes=[],
@@ -2085,6 +2094,7 @@ class CreditNotes(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.creditNotesDelete",
                 oauth2_scopes=[],