+++ accounting/src/apideck_accounting_unify/custommappings.py
@@ -85,6 +85,7 @@ class CustomMappings(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.customMappingsAll",
                 oauth2_scopes=[],
@@ -233,6 +234,7 @@ class CustomMappings(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.customMappingsAll",
                 oauth2_scopes=[],