+++ accounting/src/apideck_accounting_unify/departments.py
@@ -105,6 +105,7 @@ class Departments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsAll",
                 oauth2_scopes=[],
@@ -297,6 +298,7 @@ class Departments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsAll",
                 oauth2_scopes=[],
@@ -511,6 +513,7 @@ class Departments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsAdd",
                 oauth2_scopes=[],
@@ -700,6 +703,7 @@ class Departments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsAdd",
                 oauth2_scopes=[],
@@ -854,6 +858,7 @@ class Departments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsOne",
                 oauth2_scopes=[],
@@ -1008,6 +1013,7 @@ class Departments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsOne",
                 oauth2_scopes=[],
@@ -1200,6 +1206,7 @@ class Departments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsUpdate",
                 oauth2_scopes=[],
@@ -1392,6 +1399,7 @@ class Departments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsUpdate",
                 oauth2_scopes=[],
@@ -1543,6 +1551,7 @@ class Departments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsDelete",
                 oauth2_scopes=[],
@@ -1694,6 +1703,7 @@ class Departments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.departmentsDelete",
                 oauth2_scopes=[],