+++ accounting/src/apideck_accounting_unify/invoices.py
@@ -107,6 +107,7 @@ class Invoices(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesAll",
                 oauth2_scopes=[],
@@ -302,6 +303,7 @@ class Invoices(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesAll",
                 oauth2_scopes=[],
@@ -664,6 +666,7 @@ class Invoices(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesAdd",
                 oauth2_scopes=[],
@@ -999,6 +1002,7 @@ class Invoices(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesAdd",
                 oauth2_scopes=[],
@@ -1153,6 +1157,7 @@ class Invoices(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesOne",
                 oauth2_scopes=[],
@@ -1307,6 +1312,7 @@ class Invoices(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesOne",
                 oauth2_scopes=[],
@@ -1645,6 +1651,7 @@ class Invoices(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesUpdate",
                 oauth2_scopes=[],
@@ -1983,6 +1990,7 @@ class Invoices(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesUpdate",
                 oauth2_scopes=[],
@@ -2134,6 +2142,7 @@ class Invoices(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesDelete",
                 oauth2_scopes=[],
@@ -2285,6 +2294,7 @@ class Invoices(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.invoicesDelete",
                 oauth2_scopes=[],