+++ accounting/src/apideck_accounting_unify/models/accounting_journalentriesaddop.py
@@ -6,7 +6,7 @@ from .createjournalentryresponse import (
     CreateJournalEntryResponseTypedDict,
 )
 from .httpmetadata import HTTPMetadata, HTTPMetadataTypedDict
-from .journalentry_input import JournalEntryInput, JournalEntryInputTypedDict
+from .journalentry import JournalEntryInput, JournalEntryInputTypedDict
 from .unexpectederrorresponse import (
     UnexpectedErrorResponse,
     UnexpectedErrorResponseTypedDict,