+++ accounting/src/apideck_accounting_unify/payments.py
@@ -107,6 +107,7 @@ class Payments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsAll",
                 oauth2_scopes=[],
@@ -302,6 +303,7 @@ class Payments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsAll",
                 oauth2_scopes=[],
@@ -595,6 +597,7 @@ class Payments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsAdd",
                 oauth2_scopes=[],
@@ -861,6 +864,7 @@ class Payments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsAdd",
                 oauth2_scopes=[],
@@ -1015,6 +1019,7 @@ class Payments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsOne",
                 oauth2_scopes=[],
@@ -1169,6 +1174,7 @@ class Payments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsOne",
                 oauth2_scopes=[],
@@ -1438,6 +1444,7 @@ class Payments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsUpdate",
                 oauth2_scopes=[],
@@ -1707,6 +1714,7 @@ class Payments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsUpdate",
                 oauth2_scopes=[],
@@ -1858,6 +1866,7 @@ class Payments(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsDelete",
                 oauth2_scopes=[],
@@ -2009,6 +2018,7 @@ class Payments(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="accounting.paymentsDelete",
                 oauth2_scopes=[],