+++ accounting/src/apideck_accounting_unify/validateconnection.py
@@ -108,6 +108,7 @@ class ValidateConnection(BaseSDK):
 
         http_res = self.do_request(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.validateConnectionState",
                 oauth2_scopes=[],
@@ -279,6 +280,7 @@ class ValidateConnection(BaseSDK):
 
         http_res = await self.do_request_async(
             hook_ctx=HookContext(
+                config=self.sdk_configuration,
                 base_url=base_url or "",
                 operation_id="vault.validateConnectionState",
                 oauth2_scopes=[],